step2 refactor : domain으로 이름 변경
