+++ src/main/kotlin/blackjack/domain/CardDeck.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.domain
 
 class CardDeck private constructor(private var cards: PlayingCards) {
     fun draw(count: Int): PlayingCards {
