+++ src/main/kotlin/blackjack/domain/PlayingCards.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.domain
 
 class PlayingCards private constructor(private val cards: List<PlayingCard>) : List<PlayingCard> by cards {
     fun hasCardOf(number: CardNumber): Boolean = any { playingCard ->
