+++ src/test/kotlin/blackjack/domain/HandsTest.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
